WSG-100 rename modules
