+++ who_l3_smart_tools/core/l2/templates/concepts.cql.j2
